(QENG-1800) Install MSI for AIO package acceptance
Previously it was not possible to run Windows acceptance against FOSS
MSI packages. This commit adds a step to the aio rake tasks that can
install the puppet-agent MSI.
It depends on https://github.com/puppetlabs/beaker/pull/674

With this commit you can do:

    $ bundle exec rake ci:test:aio \
        CONFIG=config/nodes/win2012r2-rubyx64.yaml \
        SHA=<SHA> VERSION=1.0.0-o-ga354e25 \
        MSI_BASE_URL=http://int-resources.corp.puppetlabs.net/aio

Note the VERSION string must match the filename of the MSI which
includes the output from `git describe`. The host definition file needs
to specify which `ruby_arch` it wants to install, either x86 or x64.

Given one off builds may wish to be tested that are not products
of the build pipeline, the MSI url and filename may be overridden
completely with two environment variables:

* MSI_BASE_URL may optionally override the URL that Puppet can be
retrieved from, which normally defaults to:

http://builds.puppetlabs.lan/puppet(-agent)/<SHA>/artifacts/windows

* MSI_FILENAME may optionally override the filename that Puppet pulls
the MSI from, which normally defaults to:

puppet-agent-<VERSION>-<arch>.msi