(PUP-3167) Prevent regex node names from colliding
Prior to this commit, it was possible for a node resource to have
a name that collided with another resource. This is due to the fact
that when a resource is created for a node, if the node has a regex
name the regular expression will be munged and altered into a regular
string.

If the munged node name matched the name of another resource, the
two would collide and the collision would go undetected but potentially
case small errors.

In order to prevent this from happening, prepend "__node_regexp__"
to front of resource names that represent nodes which are specified
by a regex in order to prevent a possible collision.