(maint) Pin master's puppet-agent package to 0.2.1
Previously, acceptance tests using `with_puppet_running_on` would fail
when run with puppet containing AIO path changes. This was because the
`puppetserver` package installed the latest version of `puppet-agent`,
which was not compatible. As a result, the server was using an invalid
set of paths, e.g.

    :cacert "/opt/puppetlabs/agent/cache/.puppet/ssl/ca/ca_crt.pem",

This commit pins the rhel7 master to using puppet-agent 0.2.1. This way
the puppetserver and puppet-agent code are compatible (for now).

However, beaker hard codes various paths, e.g. puppetconfdir, based on
the type of acceptance, e.g. `aio`. This caused beaker to fail to backup
the ssl directory during `with_puppet_running_on`.

This commit overrides the puppet paths for the rhel7 master to use the
old locations where puppet-agent 0.2.1 expects them to be.

Ideally, beaker would always introspect the host to determine where its
puppetconfdir, etc are.