(PUP-3809) Don't create site/dist moduledirs in tests
Currently, neither the puppetserver and puppet-agent 0.3.1 packages
create the sitemoduledir, see PUP-4001. Some module related acceptance
tests explicitly created these directories, so they would pass, as well
as allow other module acceptance tests that didn't create the
directories to pass as well.

This commit moves the creation of the site and dist moduledirs to the
aio setup step until the packaging issue is resolved. It also removes
the mkdir from individual tests so that we don't have interdependent
test failures.