(PUP-3863) Update 4x functions due to API changes
Since the initial work for this pull request was done before the
changes for PUP-3978 were merged, the new 4x hiera functions were
not compatible with the changes to the way blocks were passed.

Update the new functions and corresponding specs so that they are
compatible with the changes introduced to the way 4x functions pass
blocks in PUP-3979.