(PUP-3809) Enable is_puppetserver option
The `puppet_manages_its_own_configuration_in_robust_manner` acceptance
test does not work on the JVM puppetserver, because the test deletes the
`puppet` user and group, and recreates it (with a different uid/gid). So
any files owned by the old user/group are not owned by the new
user/group.

This commit defines the `is_puppetserver` option so that the test is
properly skipped on JVM puppetserver.

Note, when the `is_puppetserver` option is enabled, beaker's defaults
its AIO path to `/etc/puppetlabs/puppetserver/conf.d`. However, the
directory currently exists in `/etc/puppetserver/conf.d` so we override
the location to match the current puppetserver build.