(PUP-4012) Dont merge lines with uppercase Q
If the fileparsing module parses a file that includes a line that starts
with an uppercase Q, this line is merged with the previous line, so

    foo
    Quest
    bar

becomes

    foouest
    bar

this can have some serious and unexpected sideeffects. The problem lies
in the way how trailing line seperators are removed. The current regex
/\n\Q/ (given the fact that line seperator is \n is most cases) looks
like some perlish regex. But in fact `\Q` does not have a special
meaning here and is translated to a literal `Q` instead. As a result the
newline and the `Q` from the following line are just removed.

However the split method does ignore trailing delimiters by default

    irb(main):003:0> "foo\nbar\n\n".split("\n")
    => ["foo", "bar"]

so I am not sure of the original intention. The fact that this caused no
issue so far is because most files (cron, /etc/hosts) that currently
make use of the provider do not use words at the start (at least not in
uppercase) but it is more likely to happen in custom providers.