(PUP-3948) Make lookup be the impl for agnostic lookup of data
This commit adds a new lookup function to the 4x API. The function
uses the same lookup semantics as the Resource, i.e. it tries to
find its data in global (hiera) scope and then in the data providers
for the environment and module.

The lookup also implements hiera-like merge semantics and will merge
the values found in the different scopes in accordance with a merge
strategy.

The extra 'merge' parameter is also added to the data provider API
to allow future enhancements there such as a data provider based
hiera replacement.

The old 3.x lookup function is removed by this commit and the its
backing class is completely rewritten.