(PUP-3632) Move rundir to /var/run/puppetlabs
Previously, the rundir defaulted to $vardir/run when running as root.

This commit changes it to /var/run/puppetlabs. The directory is
unchanged when not running as root, e.g.  ~/.puppet/var/run. Windows is
also unchanged.

This commit also modifies the git setup step to use beaker to introspect
the master host for its ssldir and pidfile settings. Note we have to
explicitly specify the master application so that we get the right
run_mode and pidfile, e.g. /var/run/puppetlabs/master.pid.