(PUP-3948) Fix issues detected in code reivew and allow undef values
This commit changes the data providers so that they throw :no_such_key
exception when the key is not found but will return nil for an existing
key associated with a nil value. The logic for the Data Binder lookup
was changed too but Hiera needs some improvement since it is currently
unable to make the distinction between a key not found and a key with a
nil value.

This commit also makes changes in the documentation for the lookup
function. Most in accordance with the comments in the code review but
some to accomodate the change in behavior.