(PUP-4018) Install and use puppetserver for AIO
Previously, the AIO setup steps would install puppet-server (redhat) or
puppetmaster-passenger (debian).

This commit modifies the setup step to install puppetserver instead, and
sets the `use-service` and `is_puppetserver` options. The `use-service`
option tells beaker to use the service scripts to bounce the service.
The `is_puppetserver` option is used by both beaker and puppet
acceptance tests to know that the JVM puppetserver is being used. Beaker
uses it to change the trapperkeeper puppetserver.conf to point to a
temporary confdir/vardir.

The `is_puppetserver` option also causes the
`puppet_manages_its_own_configuration_in_robust_manner` acceptance test
to be skipped. This is necessary because the test deletes the `puppet`
user and group, and recreates it (with a different uid/gid). So any
files owned by the old user/group are not owned by the new user/group.