(maint) Ensure puppetserver subdirectories are writable by puppet
Previously, if `puppet apply` executed before the puppetserver, then
puppet running as root could create directories under $vardir owned by
root, and not writable by the puppet group. For example:

    drwxr-x---.  2 root   puppet    6 Feb 11 12:07 bucket
    drwxr-x---.  6 root   puppet 4096 Feb 11 12:09 reports
    drwxr-x---.  2 root   puppet    6 Feb 11 12:07 server_data
    drwxr-x---.  4 puppet puppet   29 Feb 11 12:25 yaml

Later when the puppetserver tried to save a report or a filebucket file,
acceptance tests would fail, because puppetserver, running as
puppet, did not have permission to the various subdirectories. In
particular, the following acceptance tests would fail:

    tests/ssl/autosign_command.rb
    tests/resource/file/source_attribute.rb

This commit adds a temporary workaround to the AIO setup step to ensure
the above puppetserver subdirectories have permissions `puppet:puppet`
and mode 0750.

The issue of puppetserver not being able to write to its directories is
being tracked as SERVER-357. This commit should be reverted once that
issue is resolved, and is tracked in PUP-4004.