(maint) Create the puppet user and group on agents
Our acceptance tests assume the puppet user and group exist on both
master and agent. This commit adds a setup step for aio to ensure it
exists on agent-only hosts. It is not necessary to create on the master
because the puppetserver package does that.

Ticket PUP-3997 was filed to update acceptance the tests to not require
the puppet user and group, at which time this change can be reverted.