(maint) Fix hiera unit test to not fail due to coming change in hiera
One of the unit tests for the hiera function assumes that a custom backend
returns nil when a key cannot be found. This assumption makes it hard to
make a change in hiera without breaking this test in puppet.

This commit changes the test so that it performs the same task without
making this assumption.