(QENG-1800) Stop the windows service
Previously, when puppet is installed on windows, it creates and executes
a service that runs in the background. This can interfere with tests
that try to execute the agent, resulting in:

    Notice: Run of Puppet configuration client already in progress;
    skipping (C:/ProgramData/PuppetLabs/puppet/var/state/agent_catalog_run.lock
    exists)

This commit stop the service after installing the MSI.

Another approach would have been to specify
PUPPET_AGENT_STARTUP_MODE=Manual, but cygwin keeps quoting the argument
and msiexec hangs because it thinks it's a single argument as opposed to
a key=value.