(PUP-4017) Make parser an environment specific setting
Prior to this commit, the 'parser' setting was not environment
specific. In order to make it easier for users to migrate to using
the future parser, make 'parser' environment specific.

The environment parameter is optional, and if missing will use
what is configured for all environments.

This mechanism can only be used with directory environments.