(maint) Fix so that spec tests on stable can use Hiera 2.x
Two of the spec tests on stable made assumptions about how Hiera
handles 'key not found' internally. This commit removes one of
the tests since there's no way to make it function with that
assumption when using both Hiera stable and Hiera master. The
other test was tweaked to succeed with both versions.