(PUP-3090) Set always_cache_features to true for masters
While for agents caching features would not be good as they can have new
features delivered during the course of a run, for masters this poses no
problems and decreases overhead during catalog compilation. Without
this, masters will reset and reinterrogate features during the course of
operation which is expensive.