(maint) Allow puppet server installs from version
Without this change the pre-suite installs puppet-server from
nightlies/puppetserver-latest which can sometimes be an older version.
This fix allows one to specify SERVER_VERSION environment variable to
pull a specific build.
Place puppetserver-confdir option in the aio options file so that beaker
knows where the config file is prior to server install.