(PUP-2929) Optimize the TypedName for use as hash key
This commit will optimize the TypedName in two ways.

1. The initialize method now executes more than twice as fast. The
optimization consists of three improvements:
 - Use a constant string instead of a regexp when splitting the name
 - Never join the parts together. Instead use name or substring of name
 - Don't check if the name is numeric. Not needed here.

2. A compound frozen name is created when the instance is initialized.
This instance is then used when comparing and when doing to_s.

The cached @qualified attribute and associated attr_reader was
replaced with an 'qualified?' method because there's almost no
noticable gain by caching it (it's extremely fast with or without a
cache). Also, 'qualified?' (with a questionmark) is a better name.