(PUP-2929) Propagate arrays instead of splatted arguments
Micro benchmarks reveals that its much faster to pass an array
than to pass a *splat. This commit changes the interna methods
of the much used visitor to benefit from this finding. It also
changes a lot of calls that takes no additional arguments to
use the method tailored especially for this, visit_this_0.

The visit_this_0 was also optimized by now passing a NO_ARGS constant
(an empty array) instead of creating a new one on each call.