(PUP-2929) Avoid expensive deep copying when validating call params
Before this change, the method callable_PTupleType would make a
copy of the full parameter tuple when its last type was found to
be a callable (a block). This copying can be very expensive depending
on how complex the parameter types are. This commit changes the
logic to exclude the last parameter from the assignable check without
the need of a full tuple copy.