(PUP-2929) Partially revert 'Pup-909 Type inference copy optimization'
This commit is a partial revert of commit 6bf1ab4 which was an
incorrect fix to speed up the actual deep copy by replacing a
marshal/unmarshal with a more direct approach. Unfortunately, the
chosen approach was flawed. It didn't copy array or hash elements
since their content isn't reachable using instance_variables. Adding
that revealed that the marshal/unmarshal, slow as it is, was the
fastest alternative.