(PUP-1085) Be more specific about expected exception
Previously, the pacman unit test "should raise a Puppet::Error when
trying to install a group and allow_virtual is false" assumed
`allow_virtual` was false, and tried to install a package, assuming the
provider would raise a Puppet::Error.

However, in Puppet 4, `allow_virtual` is true, so the test was not
failing early as expected, and instead would actually try to execute:

    /usr/bin/pacman --noconfirm --needed --noprogressbar -Sy package'

Puppet raises Puppet::ExecutionFailure on systems without the `pacman`
executable. However, this went unnoticed on *nix, because the test was
only expecting Puppet::Error, which is the superclass for
Puppet::ExecutionFailure. This did not go unnoticed on Windows, because
puppet raises a different exception:

    Errno::ENOENT: No such file or directory - CreateProcess

This commit modifies the unit test to expect specific messages. It also
sets state directly on the resource object, e.g.
`resource[:allow_virtual] = false`. The `resource` is defined in a
`let`, so a new instance is created for each example as needed. Also
note, since `:allow_virtual` specifies `:boolean => true`, puppet
automatically creates a predicate method of the same name, e.g.
`allow_virtual?`.