(PUP-4030) Support no_proxy env variable
This change allows puppet to respect the no_proxy ENV var when set on the system.  Previously, Puppet would respect the http(s)_proxy ENV var when connecting HTTP services (e.g. puppetforge) but ignore the no_proxy setting.

Documentation is a bit inconsistent with regards to specifying domain level exception.  WGET docs for example, state that the no_proxy list should be of suffixes and thus an entry with a preceeding '.' (e.g. .example.com) creates an exception for any host under that domain.  Curl man pages and other intenet resources, suggest the use of a '*' as a wildcard.

This change supports both notations.