(maint) Change puppetserver conf.d
Previously, we had to override the location of puppetserver's conf.d to
refer to the old location and override ruby-load-path to point to where
the new puppet-agent installed puppet, facter, etc ruby code.

A recent build of puppetserver fixes the ruby-load-path issue, so that
workaround step is no longer required. Also puppetserver moved the
location of conf.d from /etc/puppetserver/conf.d to the correct location
/etc/puppetlabs/puppetserver/conf.d, so this commit updates the node
definitions to refer to the new locations.