(PUP-3296) Remove :before and :after for parameters
This commit effectively removes the `:call` option from the
`Puppet::Resource::Value` since it becomes meaningless when the only
valid settings are `:none`and `:instead`. The option is still
recognized and validated for backwards compatibility but will trigger
a deprecation warning.

The value of the `:call` option is no longer used. Instead, a check is
made if either a method or block has been given. If so, it is called.
If not, the providers setter is called.

This commit also removes a couple of @TODO's related to the `:method`
option and ensures that a given method always exists. If a block is
given, then a method is defined using that block and the name given
by the `:method` option. If not, then the name becomes an alias for
a call to the providers setter method.