(PUP-3997) Remove assumed user/group in agent tests
This commit removes the usage of `host.puppet['user']`, `host.puppet['group'],
in tests where it is used as an assumed user/group on non-master hosts.
It also removes the AIO setup workaround that ensured the puppet user and
group existed on non-master nodes.

The agent only packaging no longer creates a user and group
corresponding to the `puppet config print user` value. Tests
assuming this user/group exists on agents can no longer do so.

The previous pattern is replaced by the explicit creation of a
user on the host using the beaker `host.user_present('username')`
method.