(PUP-3997) Create site.pp with root:root permissions
Previously, the test created a `site.pp` on agent nodes without
specifying the owner and group, so the `create_test_file` method
defaulted the owner and group to `puppet`. However, the `puppet` user
isn't present on agent-only hosts.

This commit modifies the test to specify owner and group as `root`. The
test is confined to not run on Windows agents, so this is not an issue.