(PUP-4047) Rephrase error message for "non productive errors"
Before this commit, it was difficult to understand the error
messages for expressions that have no effect (only produce a value).

This changes the messages to read "This xxx has no effect", followed
by more details. This removes the need to understand the concept
"non-productive expression".

This only changes the phrasing of the error message.