(PUP-4064) Fix bogus error from 3x lookup function parameter validator
Type parameter validator, when encountering a name that was not a
string, passed the value instead of the value type to the
TypeCalculator.string method responsible for producing a string
representation of the type to include in the validation error message.
This resulted in an internal error.

This commit fixes so that the type of the value is used instead.