(PUP-3886) add yumrepo spec integration tests
Add spec integration tests to test the yumrepo type and provider.
Test include adding, changeing, deleting of yumrepo resources on the
filesystem. This change also tests a handful of properties. The full
list of properties and acceptable values are covered by spec unit tests.

This PR should supercede the previous acceptance PR.
We should still go back and add an acceptance test which ensures round
trip through yum.conf for reposdir setting and puppet resource and
setting repos in the right place in the real-world.