Initial commit for s3 file type source
I have modelled these changes on PR-3372 which is PUP-1072 in the
tickets.puppetlabs.com.

I am yet to create the spec tests and finish the documentation necessary to
formally submit it for consideration for inclusion into Puppet.

If you are interested in trying this code you will need to ensure you
have the aws-sdk-core gem installed.