(PUP-4076) Remove 'allow_virtual' deprecation warning in the 3.x series
In Puppet 4, the package type's allow_virtual setting will be true by
default. In commit 3f6d7cad, this change was made and the deprecation
warning for allow_virtual was removed in master. However, many have
found the deprecation warning to be disruptive in the 3.x series, so
this commit removes it. We will deprecate in docs instead.