(PUP-4056) Replace puppetbindir with privatebindir
Several acceptance tests attempted to execute ruby, assuming it was
within `puppetbindir`, which worked for earlier versions of AIO when
that path was `/opt/puppetlabs/puppet/bin`.

However, beaker changed `puppetbindir` to be `/opt/puppetlabs/bin`,
since that's the "public" bin directory of the puppet-agent package.
As a result, we must use `privatebindir` to resolve ruby (and openssl).

This commit modifies commands that execute on agents to use the
`ruby_command` method, which works across all supported platforms.

It modifies ruby-based ENC scripts that only execute on the master, to
resolve ruby relative to `privatebindir`.