(PUP-4056) Use privatebindir to resolve ruby on Windows
Previously, the windows_utils helper had different logic for resolving
ruby depending on whether it was testing PE or not. The PE case assumed
ruby was relative to `puppetbindir`, and everything else assumed ruby
was in the PATH. The latter is only true for git acceptance tests, but
not aio.

This commit removes the `ruby_cmd` method, and changes `profile_base` to
call `CommandUtils.ruby_command` instead. The returned command is of the
form:

    "env PATH=\"#{agent['privatebindir']}:${PATH}\" ruby"

which works for git, pe and aio type acceptance tests. It would also
work for packages type testing if that type was updated to install MSI
packages.