(PUP-3997) Default to root user/group in create_test_file
This commit alters the default user and group parameter settings
in the create_test_file method of the Puppet::Acceptance::TempFileUtils
module.

If the host is a master, the value of `puppet config print {user,group}`
are used. Otherwise, the appropriate root user and group for the host's
platform are used.