(maint) Create utility methods for ruby and gem commands
Creates module functions that return a command suitable for executing
`ruby` and `gem` across *nix and Window platforms, e.g.

    "#{ruby_command(host)} --version"

On Windows, gem is a .bat file, so must be wrapped in a `cmd /c`
shell.

Also on Windows, `privatebindir` contains multiple directories to
account for x86 vs x64 ruby running on 64-bit Windows. As a result,
`ruby_command` cannot be implemented as: `#{host['privatebindir']}/ruby`.
Instead it prepends the host's `privatebindir` to PATH in the
environment, which works for both *nix and Windows.

Note these utility methods can be executed in git, packages, pe, and aio
type acceptance tests. In the case of git & packages, `privatebindir`
refers to `/usr/bin` (ironically) so system ruby will be used. In the
case, of pe and aio, `privatebindir` refers to our vendored ruby, and we
prepend so that it takes precedence of whatever system ruby might be
installed on the system.