(PUP-3997) Run puppet apply to reset logdir owner
This commit adds a teardown step in the
tests/loader/func4x_loadable_from_modules.rb acceptance test to
run puppet apply without a puppet.conf specified in order to
restore the ownership of the logdir.

Prior to this commit, the test left ownership of the logdir set to
the user and group specified in the test, which may not be the
proper standard ownership.