(PUP-3931) Change systemd KillMode to process for agent
Currently, the entire control group is sent a kill; if a puppet agent run under the control group is running an upgrade of the puppet package, both the running agent process and the yum process doing the upgrade are killed as well as the daemon agent process, leaving the puppet installation in a bad state.  This brings the behavior in line with the behavior under the SysV init script.