(PUP-4081) lookup in module when looking up qualified names
This commit changes the data provider logic so that when a qualified
name is used in the lookup and no value is found globally or in the
environment, then the first segment of the qualified name will be used
to determine what module the lookup should be perfomed in.

No attempt will be made to do a module lookup unless the name is
qualified. This applies even if the lookup is made from within a
module.

This commit also adds a validator for the function provided module
data hash that checks that all keys are qualified with the name of
the module.