(maint) Remove unused master_url_prefix setting
When the changes were made to the Puppet 4 API to add a url prefix to all
endpoints, a `master_url_prefix` setting was added as this prefix was
initially going to be configurable. However, it was then determined that we
did not want the prefix to be configurable. While changes were made to the
code to set the prefix as a constant (see
540ec57817a7cfc6562fa725c60036210391a3d6), rendering this setting useless, the
setting was never removed. This commit removes this now-unused setting.