(PUP-4209) Introduce fail attribute in notify resource
Notify resources can output messages, but cannot alert the users about
errors or fail the current puppet run. This commit introduces the fail
attribute to notify resources, which, if set to true, will output the
message as error and fail the puppet run. The fail attribute defaults to
false and doesn't modify the current behaviour of the notify resource.