(PUP-1072) follow redirections when retrieving HTTP content
The re-implements the logic from Puppet::Network::HTTP::Connection in
a class method of Puppet::Util::HttpProxy. This is much simpler than
shoehorning proxy support into the former.

Proxy connections will use the SSL CA certificates that are trusted
by the operating system. This is generally not desirable for connections
that are made through the HTTP pool.