(PUP-4201) Fix constructor inconsistencies
This commit removes some differencies in how a Log instance was
created and its arguments validated. It also removes the ability
to change the created Log element after it has been created with
one exception. The source can be changed. Needed since the
Puppet::Reports for :log changes the source to include the name of
the host.